Update to rmem backend. Mostly hiding structs in the .c, fixes to please g++.
