+++ backends/rmem_backend.h
@@ -1,0 +1,9 @@
+#ifndef __RMEM_H__
+#define __RMEM_H__
+
+#include "rmem_generic_interface.h"
+
+rmem_layer_t* create_rmem_layer();
+
+#endif
+
